+++ src/shared/ui/Card/index.tsx
@@ -1,0 +1,1 @@
+export { Card, CardHeader, CardTitle, CardContent } from './ui'
