+++ src/shared/ui/Dialog/index.ts
@@ -1,0 +1,1 @@
+export { Dialog, DialogTrigger, DialogPortal, DialogOverlay, DialogContent, DialogHeader, DialogTitle } from './ui'
