+++ src/shared/ui/Table/index.ts
@@ -1,0 +1,1 @@
+export { Table, TableHeader, TableBody, TableRow, TableHead, TableCell } from './ui'
